refactored: user repo refactored
