+++ internal/app/user/repository/db/postgres.go
@@ -1,4 +1,4 @@
-package repository
+package db
 
 import (
 	"context"
