--- internal/app/user/repository/user_repo_test.go
@@ -1,3 +1,0 @@
-package repository
-
-//
